# backend/router/vocabulary/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Google DevJam Backend - Vocabulary API",
    "description": "Vocabulary management endpoints for creating, reading, updating, and deleting words in user's vocabulary with learning progress tracking",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/vocabulary/recommend": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "Get word recommendations",
        "description": "Get personalized word recommendations based on the user's current vocabulary. The system analyzes user's existing words, gets recommendations from Gemini AI, validates and translates new words, then stores them in the recommendation collection.",
        "operationId": "getRecommendations",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Word recommendations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendResponse"
                },
                "example": {
                  "words": [
                    {
                      "id": "1234567890123456789",
                      "word": "beautiful",
                      "translation": "美麗的",
                      "definition_zh": "美麗的",
                      "definition_en": "Having beauty; pleasing to the senses or mind",
                      "difficulty": 4,
                      "part_of_speech": "adjective",
                      "root_word": "beautiful",
                      "created_at": "2024-01-15T10:30:00Z",
                      "updated_at": "2024-01-15T10:30:00Z",
                      "learn_count": 0,
                      "fluency": 0,
                      "examples": [
                        {
                          "id": "1234567890123456790",
                          "word_id": "1234567890123456789",
                          "sentence": "She has a beautiful smile."
                        },
                        {
                          "id": "1234567890123456791",
                          "word_id": "1234567890123456789",
                          "sentence": "The sunset is beautiful tonight."
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "User not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "User not authenticated"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "user_words_error": {
                    "summary": "Failed to get user words",
                    "value": {
                      "error": "Failed to get user words: Database connection error"
                    }
                  },
                  "gemini_error": {
                    "summary": "Failed to get recommendations",
                    "value": {
                      "error": "Failed to get recommendations: Gemini API error"
                    }
                  },
                  "database_error": {
                    "summary": "Database connection error",
                    "value": {
                      "error": "Database connection error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Word": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique word identifier (Snowflake ID)",
            "example": "1234567890123456789"
          },
          "word": {
            "type": "string",
            "description": "The English word",
            "example": "hello"
          },
          "translation": {
            "type": "string",
            "description": "Traditional Chinese translation",
            "example": "你好"
          },
          "definition_zh": {
            "type": "string",
            "description": "Traditional Chinese definition",
            "example": "你好"
          },
          "definition_en": {
            "type": "string",
            "description": "English definition",
            "example": "A greeting used when meeting someone"
          },
          "difficulty": {
            "type": "integer",
            "description": "Difficulty level (1-10)",
            "minimum": 1,
            "maximum": 10,
            "example": 2
          },
          "part_of_speech": {
            "type": "string",
            "description": "Part of speech (noun, verb, adjective, etc.)",
            "example": "interjection"
          },
          "root_word": {
            "type": "string",
            "description": "Root form of the word (e.g., 'run' for 'ran')",
            "example": "hello"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Word creation timestamp",
            "example": "2024-01-15T10:30:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
      "WordExample": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique example identifier (Snowflake ID)",
            "example": "1234567890123456790"
          },
          "word_id": {
            "type": "string",
            "description": "ID of the word this example belongs to",
            "example": "1234567890123456789"
          },
          "sentence": {
            "type": "string",
            "description": "Example sentence using the word",
            "example": "Hello, how are you?"
          }
        }
      },
      "WordWithUserData": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Word"
          },
          {
            "type": "object",
            "properties": {
              "learn_count": {
                "type": "integer",
                "description": "Number of times user has practiced this word",
                "example": 5
              },
              "fluency": {
                "type": "integer",
                "description": "User's fluency level with this word (0-100)",
                "minimum": 0,
                "maximum": 100,
                "example": 80
              },
              "examples": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/WordExample"
                },
                "description": "Example sentences using this word",
                "example": [
                  {
                    "id": "1234567890123456790",
                    "word_id": "1234567890123456789",
                    "sentence": "Hello, how are you?"
                  },
                  {
                    "id": "1234567890123456791",
                    "word_id": "1234567890123456789",
                    "sentence": "Say hello to your friend."
                  }
                ]
              }
            }
          }
        ]
      },
...
      "RecommendResponse": {
        "type": "object",
        "properties": {
          "words": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WordWithUserData"
            },
            "description": "Array of recommended words with user learning data"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid request format"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from authentication endpoint"
      }
    }
  },
...
} 